requirements.txt
aiohttp==3.11.14
aioodbc==0.5.0
azure-core==1.32.0
azure-ai-language-conversations==1.1.0
azure-appconfiguration==1.7.1
azure-search-documents==11.5.2
azure-ai-textanalytics==5.3.0
azure-ai-projects==1.0.0b10
azure-identity==1.21.0
azure-cosmos==4.9.0
azure-keyvault-secrets==4.9.0
azure-mgmt-web==8.0.0
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-storage-blob==12.25.1
chardet==5.2.0
ddtrace==3.12.2
fastapi==0.115.12
langchain==0.3.23
langchain_openai==0.3.13
nltk==3.9.1
numpy==2.2.4
openai==1.74.0
opencensus-ext-azure==1.1.14
openpyxl==3.1.5
opentelemetry-sdk==1.32.1
opentelemetry-instrumentation-asgi==0.53b1
pandas==2.2.3
pre_commit==4.2.0
pydantic==2.11.3
PyJWT==2.10.1
pyodbc==5.2.0
PyYAML==6.0.2
requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
SQLAlchemy==2.0.40
starlette==0.46.2
tenacity==9.0.0
uvicorn==0.34.1
xlsxwriter==3.2.5
